requirements.txt
# NeuroFlow AI — Python dependencies
# Pin minor versions for reproducible builds; update regularly.

# ── Google ADK & AI ───────────────────────────────────────────────────────────
google-adk>=1.0.0
google-generativeai>=0.8.0
google-auth>=2.29.0
google-auth-httplib2>=0.2.0

# ── Database ──────────────────────────────────────────────────────────────────
psycopg2-binary>=2.9.9          # AlloyDB / PostgreSQL driver

# ── API Layer ─────────────────────────────────────────────────────────────────
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0

# ── HTTP client (weather tool) ────────────────────────────────────────────────
httpx>=0.27.0

# ── Utilities ─────────────────────────────────────────────────────────────────
python-dotenv>=1.0.0            # .env loading in local dev
